+++ docs/CHANGELOG.md
@@ -1,10 +1,17 @@
 Changelog
 =========
 
-1.6.3 (Unreleased)
------------------------
+
+1.7.0-beta.1 (November 7, 2022)
+-------------------------------
+- Enh #198: Implemented new HumHub v1.13 Content Form Support
+
+
+1.6.3 (November 7, 2022)
+------------------------
 - Fix #205: Hide empty list of completed tasks
 - Fix #212: Fix filter "Overdue"
+
 
 1.6.2 (October 6, 2022)
 -----------------------
@@ -31,6 +38,7 @@ Changelog
 - Enh #140: Use widget ContentVisibiltySelect
 - Fix #190: Fix wrong scheduling information for completed tasks
 - Enh #195: Remove deprecated checkbox "regular" style
+- Enh #198: Menu for creating a Task from wall stream
 
 
 1.5.5  (March 13, 2022)